Make HTTP(S) responses always preload content by default:
For most API requests (except GET object), the response is not large
and can be loaded into memory. Preloading content makes re-use of
connections automatic - the connection is returned to the connection
pool before the response object is returned to the caller of
urlopen().

The only exception, GET object, uses preload_content=False - this
allows the caller to incrementally process the response - but as a
caveat would need to release the connection manually by calling
`release_conn()` on the returned object. Docstrings have been
updated to reflect this.

Fixes #458.