fix test_list_objects_with_prefix bug when listing
Since '0' prefix was chosen it could happen such
that when listing the exact prefix match '0' due
to randomness of uuid we might have > 1 match,
this would lead to tests to fail sporadically.

This PR fixes this behavior by choosing a fixed
prefix and keeping UUID as a variable part of the
object name.