Fix date parsing
The strptime() function reads certain directives according to the
locale configured in the environment, so the parsing fails in
non-English locales, as the Last Modified date is always in
English.

This commit imports dateutil, which parses the date without using
strptime, and can correctly process English dates in any locale.

Fixes #632