adding content_hash_hex to preload_v4
In the case that the user wants to use the minio-py client to generate
a presigned url, also in the case that a server is making a request on
behalf of a command line client (where the object to be uploaded lives)
this means that the command line client will need to send the sha256sum
to the server (which is not able to calculate it). In practice I was
attempting this workflow (singularity/sregistry PR #298) and getting
an invalid signature every time. I figured out that the sha256sum was
important because it was being sent to the server from the Singularity
command line tool, and the trick was getting it into the function
preload_v4 so that it would be included in the signature calculation.
Once I was able to do this, my pre-signed urls for multipart upload
worked like a charm! While I think it would be great for minio-py to have
a fully supported function to generate these URLS more easily, in the
meantime this small change to allow specifying the content_hash_hex
will be useful to users that have similar use case to me

Signed-off-by: vsoch <vsochat@stanford.edu>