adding comment that sha256sum can be included with headers
The update to this function for presign_v4 will allow for the function
to work on behalf of a client call that wants to generate a pre-signed
url for another server, where the other server has sent the sha256sum
to the client generating the URL, and it needs to be added to the signature.
Without exposing this content_hash_hex via the headers{} dictionary,
it will not be included in the signature, and the signature calculated
by the receiving minio server (that does include it) will report
a signature mismatch. This means that if you need to include the sha256sum
in the signature, just add it to the headers dictionary.

Signed-off-by: vsoch <vsochat@stanford.edu>