Less memory allocation with put_object()
put_object() allocates more memory than what we would expect. The
reason is that workers upload parts while the main code allocates
memory from the new parts that are being read from the source stream

A semaphore is used in add_task() to make this latter blocking. That
way, the main code won't hurry to allocate new memory space for the
new data of the source stream.