Fix type annotations in remove_objects and upload_snowball_objects
The delete_object_list argument was annotated as
"Iterator[DeleteObject]". This means that mypy would complain if
client code passes a list of DeleteObject-s to remove_objects.

But in practice passing in a list works: the delete_objects function
later converts the list to an iterator using "itertools.chain".

I changed the type annotation to "Iterable[DeleteObject]". This way,
mypy is happy if we pass an iterator, and when we pass a list.

(And same thing in upload_snowball_objects)